Wrapper for Status, reorganize to avoid circular imports.
