+++ google/cloud/bigtable/table.py
@@ -28,6 +28,7 @@ from google.cloud.bigtable.column_family import _gc_rule_from_pb
 from google.cloud.bigtable.column_family import ColumnFamily
 from google.cloud.bigtable.batcher import MutationsBatcher
 from google.cloud.bigtable.batcher import FLUSH_COUNT, MAX_ROW_BYTES
+from google.cloud.bigtable.encryption import EncryptionInfo
 from google.cloud.bigtable.policy import Policy
 from google.cloud.bigtable.row import AppendRow
 from google.cloud.bigtable.row import ConditionalRow
@@ -493,7 +494,7 @@ class Table(object):
 
         :rtype: dict
         :returns: Dictionary of encryption info for this table. Keys are cluster ids and
-                  values are tuples of :class:`EncryptionInfo` instances.
+                  values are tuples of :class:`google.cloud.bigtable.encryption.EncryptionInfo` instances.
         """
         ENCRYPTION_VIEW = enums.Table.View.ENCRYPTION_VIEW
         table_client = self._instance._client.table_admin_client
@@ -1226,102 +1227,6 @@ class ClusterState(object):
         :rtype: Boolean value.
         :returns: True if  two cluster state instances are not equal.
         """
-        return not self == other
-
-
-class EncryptionInfo:
-    """Encryption information for a given resource.
-
-    If this resource is protected with customer managed encryption, the in-use Google
-    Cloud Key Management Service (KMS) key versions will be specified along with their
-    status.
-
-    :type encryption_type: int
-    :param encryption_type: See :class:`enums.EncryptionInfo.EncryptionType`
-
-    :type encryption_status: google.cloud.bigtable.table.Status
-    :param encryption_status: The encryption status.
-
-    :type kms_key_version: str
-    :param kms_key_version: The key version used for encryption.
-    """
-
-    @classmethod
-    def _from_pb(cls, info_pb):
-        return cls(
-            info_pb.encryption_type,
-            Status(info_pb.encryption_status),
-            info_pb.kms_key_version,
-        )
-
-    def __init__(self, encryption_type, encryption_status, kms_key_version):
-        self.encryption_type = encryption_type
-        self.encryption_status = encryption_status
-        self.kms_key_version = kms_key_version
-
-    def __eq__(self, other):
-        if self is other:
-            return True
-
-        if not isinstance(other, type(self)):
-            return NotImplemented
-
-        return (
-            self.encryption_type == other.encryption_type
-            and self.encryption_status == other.encryption_status
-            and self.kms_key_version == other.kms_key_version
-        )
-
-    def __ne__(self, other):
-        return not self == other
-
-
-class Status:
-    """A status, comprising a code and a message.
-
-    See: `Cloud APIs Errors <https://cloud.google.com/apis/design/errors>`_
-
-    This is a thin wrapper for ``google.rpc.status_pb2.Status``.
-
-    :type status_pb: google.rpc.status_pb2.Status
-    :param status_pb: The status protocol buffer.
-    """
-
-    def __init__(self, status_pb):
-        self.status_pb = status_pb
-
-    @property
-    def code(self):
-        """The status code.
-
-        Values are defined in ``google.rpc.code_pb2.Code``.
-
-        See: `googe.rpc.Code
-        <https://github.com/googleapis/googleapis/blob/master/google/rpc/code.proto>`_
-
-        :rtype: int
-        :returns: The status code.
-        """
-        return self.status_pb.code
-
-    @property
-    def message(self):
-        """A human readable status message.
-
-        :rypte: str
-        :returns: The status message.
-        """
-        return self.status_pb.message
-
-    def __repr__(self):
-        return repr(self.status_pb)
-
-    def __eq__(self, other):
-        if isinstance(other, type(self)):
-            return self.status_pb == other.status_pb
-        return NotImplemented
-
-    def __ne__(self, other):
         return not self == other
