+++ tests/system/test_system.py
@@ -852,6 +852,7 @@ class TestTableAdminAPI(unittest.TestCase):
 
     def test_backup(self):
         from google.cloud._helpers import _datetime_to_pb_timestamp
+        from google.cloud.bigtable import enums
 
         temp_table_id = "test-backup-table"
         temp_table = Config.INSTANCE_DATA.table(temp_table_id)
@@ -888,6 +889,10 @@ class TestTableAdminAPI(unittest.TestCase):
         self.assertEqual(temp_backup_id, temp_table_backup.backup_id)
         self.assertEqual(CLUSTER_ID_DATA, temp_table_backup.cluster)
         self.assertEqual(expire, temp_table_backup.expire_time.seconds)
+        self.assertEqual(
+            temp_table_backup.encryption_info.encryption_type,
+            enums.EncryptionInfo.EncryptionType.GOOGLE_DEFAULT_ENCRYPTION,
+        )
 
         # Testing `Backup.update_expire_time()` method
         expire += 3600  # A one-hour change in the `expire_time` parameter
